move sub-components to inline definitions
